# Conflicts:
#	app/src/main/java/com/rojasdev/apprecconproject/data/dao/RecolectoresDao.kt
#	app/src/main/res/layout/item_collecion.xml

+++ app/src/main/java/com/rojasdev/apprecconproject/data/dataModel/collectorCollection.kt
@@ -5,7 +5,7 @@ data class collectorCollection(
     val name_recolector : String,
     val PK_ID_Recoleccion: Int,
     val Cantidad : Double,
-    val Precio: Int,
+    val Precio: Double,
     val Estado: String,
     val Alimentacion: String,
     val Fecha: String,
